split parsers, and split body_html/body_text
