+++ ishar/apps/help/utils/parse/__init__.py
@@ -1,0 +1,2 @@
+from .pcls import parse_player_class
+from .stats import parse_stats
